+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use user (id) as base name for storing files
 - Rename ID -> Transcript
 - Remove the "exon" part from the exon string on display
+- Use createObjectURL to download files with JWT
+
+### Removed
+- Removed unauthenticated access in debug config
 
 ## 0.0.2